Revert "Factor out captured variables that are not reused outside findGit callback"
This reverts commit df005bdda519c63366cd3590a83a10285d7a4b57.
